osc: proper handling of fullscreen state dependent visibility events
Fixes:
* reset margins when using boxvideo with showfullscreen=no
* visibility can now be changed in fs when using showfullscreen=no
* set_osd is no longer called every tick() when the OSC is not visible

This fixes an issue where using boxvideo with showfullscreen=no or
showwindowed=no resulted in the margins not resetting when the OSC
wasn't visible.

For example, using:

script-opts=osc-visibility=always,osc-boxvideo=yes,osc-showfullscreen=no

and then going fullscreen would make the OSC disappear but the video
margins would remain. This is because the margin conditionals weren't
dependent on showfullscreen/showwindowed. Instead, if you specified
these options, the OSC was cleared using set_osd on tick() but
state.enabled and state.osc_visible would remain true which meant the
margins wouldn't reset. Subsequently calling set_osd on every tick() is
somewhat inefficient if the OSC is not visible anyway.

To fix this, showfullscreen/showwindowed are now directly tied to
fullscreen state changes which trigger the visibility_mode function when
the right conditions are met. Using visibility_mode correctly sets
state.enabled and state.osc_visible as well as calls update_margins() so
both the visibility and the margins are automatically handled properly.

Runtime visibility changes are now also allowed when no-show options are
specified. Previously if showfullscreen or showwindowed were set to
false it was not possible to toggle or cycle the visibility in a
non-rendering state (e.g. when fullscreen + showfullscreen=no).

state.prev_vis is set when the OSC is disabled by fullscreen state
changes in order to track the previous visibility mode, allowing memory
of it for when the video returns to the previous fullscreen state.
If the user sets the OSC to a visible mode before returning to this
previous state then the player will use the current visibility mode
instead of the previous one. state.prev_vis is reset outside of pending
a return to the previous state in order to prevent edge case issues with
the options changing at runtime.